Move `PartitionEvaluator` and window_state structures to `datafusion_expr` crate (#6690)
* Move `PartitonEvaluator` and window_state structures to `datafusion_expr` crate

* Update docs
