+++ datafusion/expr/src/lib.rs
@@ -41,6 +41,7 @@ mod literal;
 pub mod logical_plan;
 mod nullif;
 mod operator;
+mod partition_evaluator;
 mod signature;
 pub mod struct_expressions;
 mod table_source;
@@ -51,6 +52,7 @@ mod udf;
 pub mod utils;
 pub mod window_frame;
 pub mod window_function;
+pub mod window_state;
 
 pub use accumulator::Accumulator;
 pub use aggregate_function::AggregateFunction;
@@ -69,6 +71,7 @@ pub use literal::{lit, lit_timestamp_nano, Literal, TimestampLiteral};
 pub use logical_plan::*;
 pub use nullif::SUPPORTED_NULLIF_TYPES;
 pub use operator::Operator;
+pub use partition_evaluator::PartitionEvaluator;
 pub use signature::{Signature, TypeSignature, Volatility};
 pub use table_source::{TableProviderFilterPushDown, TableSource, TableType};
 pub use udaf::AggregateUDF;
